requirements.txt
flask
google-genai
markdown